requirements.txt
torch~=2.5.1
numpy==2.2.1
pytorch-lightning~=2.5.0.post0
invoke~=2.2.0
torchmetrics~=1.6.1
omegaconf~=2.3.0
hydra-core~=1.3.2